2.2.7.86
Added Foundry support
Added Charcoal Pit support
Added Static Power support
Added Minestrappolation support
Added Buildcraft Ore Processing support
Added Dense Plates
Made Tiny Dusts and Small Dusts their own module
Added pt_BR.lang
Updated ja_JP.lang
Made AbyssalCraft items' information locked
Added utility method to get all items for an OreDict entry
Added utility method to make an EnumSet

Changed utils' item and fluid cache to actual caches
Changed AE2 and FuturePack module to not use reflection

+++ src/main/java/thelm/jaopca/modules/ModuleExNihiloCreatio.java
@@ -52,7 +52,7 @@ public class ModuleExNihiloCreatio extends ModuleBase {
 			"Coal", "Lapis", "Diamond", "Emerald", "Quartz", "QuartzBlack", "CertusQuartz", "Redstone", "Glowstone"
 			);
 
-	public static final HashMap<IOreEntry, ItemStack> ORE_SOURCES = Maps.<IOreEntry, ItemStack>newHashMap(); 
+	public static final HashMap<IOreEntry, ItemStack> ORE_SOURCES = Maps.<IOreEntry, ItemStack>newHashMap();
 	public static final HashMap<IOreEntry, float[]> SIEVE_CHANCES = Maps.<IOreEntry, float[]>newHashMap();
 
 	@Override
@@ -207,7 +207,7 @@ public class ModuleExNihiloCreatio extends ModuleBase {
 			return chance / entry.getRarity();
 		}
 		else {
-			return chance / entry.getRarity() / 100D * (double)ModConfig.world.normalDropPercent;
+			return chance / entry.getRarity() / 100D * ModConfig.world.normalDropPercent;
 		}
 	}
